removes console.log, adds umd, refactors api
This commit removes all of the console.logs that we had around. It also sets the default behavior of the lib to keep searching for qrcodes after it encountered one, and not just search once - in the next commit searching once will be enabled by an argument.
The method setDecoderCallback is removed as we now set the result callback from the prepareVideo (and prepareImage later).
There is also a bit of refactor to package.json and gulpfile for a better build process.

fixes #14